Remove BOARD_SYSTEMSDK_VERSIONS
The value automatically set by the build system is more correct. Setting
it manually to 33 actually was erronous because there are some modules
installed to product or vendor partition which have `sdk_version` set to
`system_current` which is VIC as of now. Those errors were not detected
due to a bug in the build system, but that is now detected after the
refactoring done as part of https://android-review.git.corp.google.com/q/topic:%22limit_systemsdk%22

Bug: 318658392
Bug: 320732555
Change-Id: Ic836751e4b689216fe7f1fb8032542ecf1b96308
(cherry picked from commit e87b6977960867cbcf45b5be2507daf09ac94265)